Remove all UI and start login from ground up
