+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@apollo/react-hoc": "^3.1.3",
     "@apollo/react-hooks": "^3.1.1",
     "@craco/craco": "^5.5.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.25",
@@ -22,6 +23,7 @@
     "@types/reselect": "^2.2.0",
     "@xstate/react": "^0.7.1",
     "apollo-boost": "^0.4.4",
+    "apollo-cache-inmemory": "^1.6.3",
     "apollo-server": "^2.9.5",
     "apollo-server-express": "^2.9.4",
     "bcryptjs": "^2.4.3",
