+++ src/App.tsx
@@ -7,25 +7,26 @@ import { Routes } from "./common/routes";
 import AuthRoute from "./components/AuthRoute";
 import { useCampaignState } from "./context/campaign/store";
 import CampaignSelect from "./pages/CampaignSelect";
-import RefactoredJournal from "./pages/RefactoredJournal";
+import Journal from "./pages/Journal";
 
 const App: React.FC = () => {
   const { activeCampaign } = useCampaignState();
 
   return (
     <div className="App">
-      <Switch>
+      ts
+      {/* <Switch>
         <AuthRoute
           path={Routes.JOURNAL}
           isAuth={Boolean(activeCampaign)}
-          component={RefactoredJournal}
+          component={Journal}
           redirectUrl={Routes.CAMPAIGN_SELECT}
         />
         <Route path={Routes.CAMPAIGN_SELECT} component={CampaignSelect} />
         <Redirect
           to={activeCampaign ? Routes.JOURNAL : Routes.CAMPAIGN_SELECT}
         />
-      </Switch>
+      </Switch> */}
     </div>
   );
 };
