+++ src/api/client.ts
@@ -1,0 +1,20 @@
+import ApolloClient, { InMemoryCache } from "apollo-boost";
+
+import { LocalStorage } from "../common/constants";
+import { valueFromStorage } from "../common/helpers";
+
+export const client = new ApolloClient({
+  uri: "http://localhost:4000",
+  cache: new InMemoryCache(),
+  resolvers: {
+
+  },
+  request: operation => {
+    const token = valueFromStorage(LocalStorage.TOKEN);
+    operation.setContext({
+      headers: {
+        authorization: token ? `Bearer ${token}` : ""
+      }
+    });
+  }
+});
