+++ src/api/enhancers.ts
@@ -1,0 +1,4 @@
+import { graphql } from "@apollo/react-hoc";
+import { CAMPAIGNS } from "./gqls";
+
+export const withCampaigns = graphql(CAMPAIGNS);
