+++ src/pages/Journal/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Journal = () => {
+  return <div>Journal</div>;
+};
+
+export default Journal;
