+++ src/pages/Login/LoginLayout.tsx
@@ -1,0 +1,3 @@
+import React from "react"
+
+const LoginLayout = () => <div></div>
