--- src/pages/RefactoredJournal/index.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const RefactoredJournal = () => {
-  return <div>Journal</div>;
-};
-
-export default RefactoredJournal;
